fix(apple-music): broken list items with comma
When we had track, album or playlist items with a comma the regex used
for creating table removed necessary parts of the string. This commit
fixes this issue by using osascript's `-s` flag with a value `s` which
returns a recompilable source form. The new type of return then can be
processed into a table with `loadstring`.